syntax fixes and buildifier and sanitization
