+++ tensorboard/webapp/plugins/BUILD
@@ -21,7 +21,7 @@ ng_module(
         "//tensorboard/webapp/core",
         "//tensorboard/webapp/core/store",
         "//tensorboard/webapp/plugins/npmi",
-        "//tensorboard/webapp/plugins/text_v2/tf_text_v2:text_v2",
+        "//tensorboard/webapp/plugins/text_v2",
         "//tensorboard/webapp/types",
         "@npm//@angular/common",
         "@npm//@angular/core",
