+++ tensorboard/webapp/plugins/text_v2/effects/BUILD
@@ -6,14 +6,14 @@ package(default_visibility = ["//tensorboard:internal"])
 ng_module(
     name = "effects",
     srcs = [
-        "text_effects.ts",
         "index.ts",
+        "text_effects.ts",
     ],
     deps = [
-        "//tensorboard/webapp/plugins/text_v2/tf_text_v2/actions",
-        "//tensorboard/webapp/plugins/text_v2/tf_text_v2/data_source",
-        "//tensorboard/webapp/plugins/text_v2/tf_text_v2/store:types",
         "//tensorboard/webapp/core/actions",
+        "//tensorboard/webapp/plugins/text_v2/actions",
+        "//tensorboard/webapp/plugins/text_v2/data_source",
+        "//tensorboard/webapp/plugins/text_v2/store:types",
         "@npm//@angular/common",
         "@npm//@angular/core",
         "@npm//@ngrx/effects",
@@ -21,4 +21,3 @@ ng_module(
         "@npm//rxjs",
     ],
 )
-
