+++ tensorboard/webapp/plugins/text_v2/effects/index.ts
