+++ tensorboard/webapp/plugins/text_v2/store/text_types.ts
