+++ tensorboard/webapp/plugins/text_v2/views/step/step_module.ts
@@ -12,11 +12,17 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
+
+import {CommonModule} from '@angular/common';
 import {NgModule} from '@angular/core';
-import {Tftext2HttpServerDataSource} from './tftext2_data_source';
-import {TBHttpClientModule} from '../../../../webapp_data_source/tb_http_client_module';
+import {MatCardModule} from '@angular/material/card';
+
+import {MarkdownRendererModule} from '../../../../widgets/markdown_renderer/markdown_renderer_module';
+import {StepComponent} from './step_component';
+
 @NgModule({
-  imports: [TBHttpClientModule],
-  providers: [Tftext2HttpServerDataSource],
+  declarations: [StepComponent],
+  imports: [CommonModule, MatCardModule, MarkdownRendererModule],
+  exports: [StepComponent],
 })
-export class Tftext2ServerDataSourceModule {}
+export class StepModule {}
