+++ tensorboard/webapp/plugins/text_v2/views/tag/tag.css
@@ -1,3 +1,7 @@
+.tag-container {
+  margin-bottom: 5px;
+}
+
 .step-list {
   border-radius: 3px;
   border: 2px solid;
