+++ tensorboard/webapp/plugins/text_v2/views/tag/tag_component.ts
