+++ tensorboard/webapp/plugins/text_v2/views/tag_group/tag_group_component.ts
